Separate naive and cuzk
