+++ src/cuzk/shader_manager.rs
@@ -9,9 +9,7 @@ pub static DECOMPOSE_SCALARS_SHADER: Lazy<String> = Lazy::new(|| {
 pub static EXTRACT_WORD_FROM_BYTES_LE_FUNCS: Lazy<String> = Lazy::new(|| {
     include_str!("wgsl/cuzk/extract_word_from_bytes_le.template.wgsl").to_string()
 });
-pub static MONTGOMERY_PRODUCT_FUNCS_2: Lazy<String> = Lazy::new(|| {
-    include_str!("wgsl/montgomery/mont_product.template.wgsl").to_string()
-});
+
 pub static MONTGOMERY_PRODUCT_FUNCS: Lazy<String> = Lazy::new(|| {
     include_str!("wgsl/montgomery/mont_pro_product.template.wgsl").to_string()
 });
@@ -251,7 +249,6 @@ impl ShaderManager {
         handlebars.register_template_string("bigint_funcs", BIGINT_FUNCS.as_str()).unwrap();
         handlebars.register_template_string("field_funcs", FIELD_FUNCS.as_str()).unwrap();
         handlebars.register_template_string("montgomery_product_funcs", MONTGOMERY_PRODUCT_FUNCS.as_str()).unwrap();
-        handlebars.register_template_string("montgomery_product_funcs_2", MONTGOMERY_PRODUCT_FUNCS_2.as_str()).unwrap();
         handlebars.register_template_string("barrett_funcs", BARRETT_FUNCS.as_str()).unwrap();
 
         let data = json!({
@@ -271,37 +268,6 @@ impl ShaderManager {
             "rinv_limbs": self.rinv_limbs,
         });
         handlebars.render("test_field", &data).unwrap()
-    }
-
-    pub fn gen_test_diff_impl_shader(&self) -> String {
-        let mut handlebars = Handlebars::new();
-        handlebars.register_template_string("test_diff_impl", TEST_DIFF_IMPL_SHADER.as_str()).unwrap();
-        
-
-        handlebars.register_template_string("structs", STRUCTS.as_str()).unwrap();
-        handlebars.register_template_string("bigint_funcs", BIGINT_FUNCS.as_str()).unwrap();
-        handlebars.register_template_string("field_funcs", FIELD_FUNCS.as_str()).unwrap();
-        handlebars.register_template_string("montgomery_product_funcs", MONTGOMERY_PRODUCT_FUNCS.as_str()).unwrap();
-        handlebars.register_template_string("montgomery_product_funcs_2", MONTGOMERY_PRODUCT_FUNCS_2.as_str()).unwrap();
-        handlebars.register_template_string("barrett_funcs", BARRETT_FUNCS.as_str()).unwrap();
-
-        let data = json!({
-            "word_size": self.word_size,
-            "num_words": self.num_words,
-            "p_limbs": self.p_limbs,
-            "p_limbs_plus_one": self.p_limbs_plus_one,
-            "zero_limbs": self.zero_limbs,
-            "one_limbs": self.one_limbs,
-            "r_limbs": self.r_limbs,
-            "w_mask": self.w_mask,
-            "num_words_mul_two": self.num_words * 2,
-            "num_words_plus_one": self.num_words + 1,
-            "n0": self.n0,
-            "mu_limbs": self.mu_limbs,
-            "slack": self.slack,
-            "rinv_limbs": self.rinv_limbs,
-        });
-        handlebars.render("test_diff_impl", &data).unwrap()
     }
 
     pub fn gen_test_point_shader(&self) -> String {
