--- src/cuzk/test/diff_impl.rs
@@ -1,117 +1,0 @@
-use std::time::Instant;
-
-use ff::PrimeField;
-use wgpu::CommandEncoderDescriptor;
-
-use crate::{cuzk::{
-    gpu::{
-        create_and_write_storage_buffer, create_bind_group, create_bind_group_layout,
-        create_compute_pipeline, create_storage_buffer, execute_pipeline, get_adapter, get_device,
-        read_from_gpu, read_from_gpu_test,
-    },
-    msm::{PARAMS, WORD_SIZE},
-    shader_manager::ShaderManager,
-    utils::{field_to_u8_vec_for_gpu, field_to_u8_vec_montgomery_for_gpu, u8s_to_field_without_assertion},
-}, utils::montgomery::field_to_bytes_montgomery};
-
-pub async fn field_op<F: PrimeField>(op: &str, a: F, b: F) -> (F, F) {
-    let a_bytes = field_to_u8_vec_montgomery_for_gpu(&a, PARAMS.num_words, WORD_SIZE);
-    println!("A bytes: {:?}", a_bytes);
-    let b_bytes = field_to_u8_vec_montgomery_for_gpu(&b, PARAMS.num_words, WORD_SIZE);
-    println!("B bytes: {:?}", b_bytes);
-
-    // let t = field_to_bytes_montgomery(&a);
-    // println!("T bytes: {:?}", t);
-    let input_size = 1;
-    let chunk_size = if input_size >= 65536 { 16 } else { 4 };
-    let num_words = PARAMS.num_words;
-    println!("Input size: {}", input_size);
-    println!("Chunk size: {}", chunk_size);
-    println!("Num words: {}", num_words);
-    println!("Word size: {}", WORD_SIZE);
-    println!("Params: {:?}", PARAMS);
-
-    let shader_manager = ShaderManager::new(WORD_SIZE, chunk_size, input_size);
-
-    let adapter = get_adapter().await;
-    let (device, queue) = get_device(&adapter).await;
-    let mut encoder = device.create_command_encoder(&CommandEncoderDescriptor {
-        label: Some("Field Encoder"),
-    });
-
-    let shader_code = shader_manager.gen_test_diff_impl_shader();
-
-    let a_sb = create_and_write_storage_buffer(Some("A buffer"), &device, &a_bytes);
-    let b_sb = create_and_write_storage_buffer(Some("B buffer"), &device, &b_bytes);
-
-    let result_1_sb = create_storage_buffer(Some("Result buffer 1"), &device, (num_words * 4) as u64);
-    let result_2_sb = create_storage_buffer(Some("Result buffer 2"), &device, (num_words * 4) as u64);
-
-    let bind_group_layout = create_bind_group_layout(
-        Some("Bind group layout"),
-        &device,
-        vec![],
-        vec![&a_sb, &b_sb, &result_1_sb, &result_2_sb],
-        vec![],
-    );
-
-    let bind_group = create_bind_group(
-        Some("Bind group"),
-        &device,
-        &bind_group_layout,
-        vec![&a_sb, &b_sb, &result_1_sb, &result_2_sb],
-    );
-
-    let compute_pipeline = create_compute_pipeline(
-        Some("Field add shader"),
-        &device,
-        &bind_group_layout,
-        &shader_code,
-        op,
-    )
-    .await;
-
-    execute_pipeline(&mut encoder, compute_pipeline, bind_group, 1, 1, 1).await;
-
-    // Map results back from GPU to CPU.
-    let data = read_from_gpu_test(&device, &queue, encoder, vec![result_1_sb, result_2_sb]).await;
-
-    // Destroy the GPU device object.
-    device.destroy();
-
-
-    let result_1 = u8s_to_field_without_assertion::<F>(&data[0], num_words, WORD_SIZE);
-    let result_2 = u8s_to_field_without_assertion::<F>(&data[1], num_words, WORD_SIZE);
-
-    (result_1, result_2)
-}
-
-pub fn run_webgpu_field_op<F: PrimeField>(op: &str, a: F, b: F) -> (F, F) {
-    pollster::block_on(run_webgpu_field_op_async(op, a, b))
-}
-
-pub async fn run_webgpu_field_op_async<F: PrimeField>(op: &str, a: F, b: F) -> (F, F) {
-    let now = Instant::now();
-    let result = field_op::<F>(op, a, b).await;
-    println!("Field add time: {:?}", now.elapsed());
-    result
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ff::Field;
-    use halo2curves::bn256::Fq;
-    use rand::thread_rng;
-
-    #[test]
-    fn test_webgpu_diff_impl() {
-        let mut rng = thread_rng();
-        let a = Fq::random(&mut rng);
-        let b = Fq::random(&mut rng);
-
-        let (result_1, result_2) = run_webgpu_field_op::<Fq>("test_diff_impl", a, b);
-
-        assert_eq!(result_1, result_2);
-    }
-}
