--- src/halo2curves/test/mod.rs
@@ -1,2 +1,0 @@
-pub mod pallas;
-pub mod bn256;
