+++ src/lib.rs
@@ -1,4 +1,2 @@
-pub mod gpu;
-pub mod utils;
-pub mod halo2curves;
+pub mod naive;
 pub mod cuzk;
