+++ src/naive/halo2curves/test/mod.rs
@@ -1,0 +1,2 @@
+pub mod bn256;
+pub mod pallas;
