+++ src/naive/wgsl/main.wgsl
