+++ src/naive/wgsl/storage.wgsl
