+++ src/naive/wgsl/test/constants.wgsl
