+++ src/naive/wgsl/test/point.wgsl
