+++ src/naive/wgsl/test/sum_of_sums.wgsl
